requirements.txt
flask==2.3.2
flask-cors==3.0.10
gunicorn==20.1.0
joblib==1.3.2
numpy==1.24.3
pandas==2.0.3
psycopg2-binary==2.9.7
python-dotenv==1.0.0
requests==2.31.0
scikit-learn==1.3.0
xgboost==1.7.6
matplotlib==3.7.1
scipy==1.10.1
ollama
